requirements.txt
colorama==0.4.4
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
timm==0.4.12
torch==1.7.1+cu101
torchvision==0.8.2+cu101
tqdm==4.62.3
typing-extensions==3.10.0.2
